+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to BongoCat-mac will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.3] - 2025-08-07
+
+### Modified
+
+- Update `push.sh` to include zipping the pkg file before upload.
+- Update `run.sh` to include zipping the pkg file before upload.
+
 ## [1.8.2] - 2025-08-07
 
 ### Added